requirements.txt
python-telegram-bot==13.7
gspread==3.7.0
oauth2client==4.1.3
google-api-python-client==2.36.0
Flask==2.0.2
pytz==2021.3
Pillow==9.0.0
python-dotenv==0.19.0